Module Sihl_queue.MakePolling

Parameters

ScheduleService : Sihl.Schedule.Sig.SERVICE

Repo : Sig.REPO

Signature

include Sig.Core.Container.Service.Sig

val lifecycle : Sihl_core.Container.Lifecycle.t

val dispatch : Sig.Core.Ctx.t -> job:'a Sig.Job.t -> ?⁠delay:Sig.Utils.Time.duration -> 'a -> unit Lwt.t

Queue a job for processing. Use delay to run the initially job after a certain amount of time.

val register_jobs : Sig.Core.Ctx.t -> jobs:'a Sig.Job.t list -> unit Lwt.t

Register jobs that can be dispatched.

Only registered jobs can be dispatched. Dispatching a job that was not registered does nothing.

val register : ?⁠jobs:'a Sig.Job.t list -> unit -> Sig.Core.Container.Service.t